feat006_backend/infer/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
websockets==12.0
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5
